+++ CSM Security Database Core/CHANGELOG.md
@@ -1,5 +1,18 @@
 ﻿# CSM Security Database CHANGELOG
 
+## [1.0.1] - 25.12-2025
+
+### Patch
+
+- Added extension to subscribe automatically all security database services required.
+
+#### Dependencies
+
+| Package                                 | Previous Version | New Version     |
+|:----------------------------------------|:----------------:|:---------------:|
+| CSM.Database.Core                       | 4.1.2            | 4.1.2           |
+| Microsoft.EntityFrameworkCore.Design	  | 10.0.1           | 10.0.1          |
+
 ## [1.0.0] - 24.12-2025
 
 ### Init